chore: 编译警告清零（主工程/GUI/MAUI 三端 0 警告）
- 主工程 PackFile CS0675 位运算符号扩展 → (uint) 强转
- GUI 9 个 CS8602 空引用：Avalonia TextBlock.Inlines nullable → tb.Inlines!.Add
- MAUI 真实修复：TappedEventArgs? 参数、PickPhotoAsync→PickPhotosAsync、
  custom.DefaultBaseUrl! nullable 元组
- MAUI 128 个注释/平台噪音（CS1570/CS1587 XML 注释、CA1416/CA1422 Android
  Console API、CA2255 ModuleInitializer）→ csproj NoWarn（共享桌面代码特有）
- 三端 0 警告 0 错误；自测 4793 全过

+++ WayCoder/Git/PackFile.cs
@@ -453,7 +453,7 @@ public static class PackFileReader
             while ((c & 0x80) != 0)
             {
                 c = pack[pos++];
-                off = ((off + 1) << 7) | (c & 0x7F);
+                off = ((off + 1) << 7) | (uint)(c & 0x7F); // CS0675：c&0x7F 是 int，显式转无符号避免符号扩展
             }
             baseOffset = offset - off;
         }
